iota_config

Module genesis

Structs

Delegation

Represents a delegation of stake and gas payment to a validator, coming from a delegator. This struct is used to serialize and deserialize delegations to and from a csv file.

Delegations

Represents genesis delegations to validators.

Genesis

GenesisCeremonyParameters

Initial set of parameters for a chain.

GenesisChainParameters

TokenAllocation

TokenDistributionSchedule

TokenDistributionScheduleBuilder

UnsignedGenesis

ValidatorAllocation

Represents the allocation of stake and gas payment to a validator.

Functions

csv_reader_with_comments

Helper function to create a CSV reader with custom settings. In this case, it sets the comment character to #.