iota_config

Module node

Structs

ArchiveReaderConfig

AuthorityKeyPairWithPath

Wrapper struct for AuthorityKeyPair that can be deserialized from a file path.

AuthorityOverloadConfig

Configuration for the threshold(s) at which we consider the system to be overloaded. When one of the threshold is passed, the node may stop processing new transactions and/or certificates until the congestion resolves.

AuthorityStorePruningConfig

CheckpointExecutorConfig

ConsensusConfig

DBCheckpointConfig

ExecutionCacheConfig

ExpensiveSafetyCheckConfig

Genesis

GrpcApiConfig

Configuration for the gRPC API service

KeyPairWithPath

Wrapper struct for IotaKeyPair that can be deserialized from a file path. Used by network, worker, and account keypair.

MetricsConfig

NodeConfig

StateArchiveConfig

StateDebugDumpConfig

Configurations which determine how we dump state debug info. Debug info is dumped when a node forks.

StateSnapshotConfig

TransactionKeyValueStoreReadConfig

TransactionKeyValueStoreWriteConfig

WritebackCacheConfig

Enums

ConsensusProtocol

ExecutionCacheType

RunWithRange

ServerType

Constants

DEFAULT_COMMISSION_RATE

Default commission rate of 2%

DEFAULT_GRPC_CONCURRENCY_LIMIT

DEFAULT_VALIDATOR_GAS_PRICE

Default gas price of 1000 Nanos

Functions

bool_true

default_admin_interface_address

default_concurrency_limit

default_enable_index_processing

default_end_of_epoch_broadcast_channel_capacity

default_grpc_api_config

default_json_rpc_address

default_zklogin_oauth_providers

Type Aliases

ExecutionCacheTypeAtomicU8

Type alias for atomic representation of ExecutionCacheType for lock-free operations